// src/codegen/api/product/models/archive-filter-schema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ArchiveFilterOptionSchema } from "./archive-filter-option-schema.ts";

export const archiveFilterSchemaSlugEnum = {
  entity: "entity",
  sort: "sort",
  templateIds: "templateIds",
  profileIds: "profileIds",
  toneIds: "toneIds",
  videoTypeIds: "videoTypeIds",
  platformIds: "platformIds",
  videoDurationIds: "videoDurationIds",
} as const;

export type ArchiveFilterSchemaSlugEnumKey =
  (typeof archiveFilterSchemaSlugEnum)[keyof typeof archiveFilterSchemaSlugEnum];

export const archiveFilterSchemaTypeEnum = {
  select: "select",
  multiselect: "multiselect",
} as const;

export type ArchiveFilterSchemaTypeEnumKey =
  (typeof archiveFilterSchemaTypeEnum)[keyof typeof archiveFilterSchemaTypeEnum];

/**
 * Archive filter
 * @description Archive filter description
 */
export type ArchiveFilterSchema = {
  /**
   * @type string
   */
  slug: ArchiveFilterSchemaSlugEnumKey;
  /**
   * @type string
   */
  name: string;
  icon: string | null;
  /**
   * @type string
   */
  type: ArchiveFilterSchemaTypeEnumKey;
  /**
   * @type array
   */
  options: ArchiveFilterOptionSchema[];
};